[StyleCop] Fix new warnings in Test TestHelpers - Add files for classes: RecognizerExtensions, TestModelExtensions, TestResources, TestResourcesExtensions, TestUtils - Add ruleset to the project - Document enums - Add trailing comma - Add and remove spacing when needed
