+++ .NET/Microsoft.Recognizers.Text.DataDrivenTests/TestUtils.cs
@@ -1,0 +1,98 @@
+﻿using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Microsoft.Recognizers.Text.Choice;
+using Microsoft.Recognizers.Text.DateTime;
+using Microsoft.Recognizers.Text.DateTime.English;
+using Microsoft.Recognizers.Text.DateTime.French;
+using Microsoft.Recognizers.Text.DateTime.German;
+using Microsoft.Recognizers.Text.DateTime.Italian;
+using Microsoft.Recognizers.Text.DateTime.Portuguese;
+using Microsoft.Recognizers.Text.DateTime.Spanish;
+using Microsoft.Recognizers.Text.Number;
+using Microsoft.Recognizers.Text.NumberWithUnit;
+using Microsoft.Recognizers.Text.Sequence;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Newtonsoft.Json;
+using DateObject = System.DateTime;
+
+namespace Microsoft.Recognizers.Text.DataDrivenTests
+{
+
+    public static class TestUtils
+    {
+        public static string GetCulture(string source)
+        {
+            var langStr = source.Substring(source.LastIndexOf('_') + 1);
+            return Culture.SupportedCultures.First(c => c.CultureName == langStr).CultureCode;
+        }
+
+        public static bool EvaluateSpec(TestModel spec, out string message)
+        {
+            if (string.IsNullOrEmpty(spec.Input))
+            {
+                message = $"spec not found";
+                return true;
+            }
+
+            if (spec.IsNotSupported())
+            {
+                message = $"input '{spec.Input}' not supported";
+                return true;
+            }
+
+            if (spec.IsNotSupportedByDesign())
+            {
+                message = $"input '{spec.Input}' not supported by design";
+                return true;
+            }
+
+            message = string.Empty;
+
+            return false;
+        }
+
+        public static string SanitizeSourceName(string source)
+        {
+            return source.Replace("Model", "").Replace("Extractor", "").Replace("Parser", "");
+        }
+
+        public static Models GetModel(string source)
+        {
+            var model = SanitizeSourceName(source);
+            Models modelEnum = Models.Number;
+            if (Enum.TryParse(model, out modelEnum))
+            {
+                return modelEnum;
+            }
+
+            throw new Exception($"Model '{model}' not supported");
+        }
+
+        public static DateTimeParsers GetParser(string source)
+        {
+            var parser = SanitizeSourceName(source);
+            DateTimeParsers parserEnum = DateTimeParsers.Date;
+            if (Enum.TryParse(parser, out parserEnum))
+            {
+                return parserEnum;
+            }
+
+            throw new Exception($"Parser '{parser}' not supported");
+        }
+
+        public static DateTimeExtractors GetExtractor(string source)
+        {
+            var extractor = SanitizeSourceName(source);
+            DateTimeExtractors extractorEnum = DateTimeExtractors.Date;
+            if (Enum.TryParse(extractor, out extractorEnum))
+            {
+                return extractorEnum;
+            }
+
+            throw new Exception($"Extractor '{extractor}' not supported");
+        }
+    }
+}
